Setting module refactored to Config module approach
