+++ src/nip33.rs
@@ -1,7 +1,5 @@
+use crate::config::settings::Settings;
 use crate::lightning::LnStatus;
-use crate::Settings;
-use crate::LN_STATUS;
-use crate::Settings;
 use crate::LN_STATUS;
 use chrono::Duration;
 use mostro_core::error::{MostroError, MostroError::MostroInternalErr};
